Moving side ways to split game and platform.
I am angling the code to split better split the game and platform
layers. Ultimate goal is to have the game be totally separate that I can
load in a dll. All I have now is a call to game update and init game so
I am just where I want to be.

Still need to work out how to do sound a bit better but it works for
now. I removed the two bools I was using to trigger the sound effects
and instead pass a single pointer which just plays something if the
pointer isn't null.

Added some debug methods for doing temporary resource loading so now I
can load the bmps and waves in the game layer.