Hot reloading game code!
Ambition realized! I've set up hot reloading for the game's dll. I then
put it to great use while hacking in some basic game states for a
pausing and a main menu. It is hard to believe how well it actually
worked worked. The code is starting to look a bit grimy though so
probably time to start scrubbing it down a bit.